pyproject.toml
[project]
name = "ai-tg-bot-orchestrator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "aiokafka>=0.13.0",
    "company-logger",
    "fastapi>=0.128.0",
    "google-genai>=1.57.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "ruff>=0.14.10",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4

target-version = "py39"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"

indent-style = "space"

skip-magic-trailing-comma = false

line-ending = "auto"

docstring-code-format = false

docstring-code-line-length = "dynamic"

[tool.uv.sources]
company-logger = { git = "ssh://git@github.com/TereshenkoM/ai_tg_bot_logger.git", rev = "v0.1.0" }
